card-gen/requirements.txt
google_api_python_client==2.68.0
google_auth_oauthlib==0.7.1
Pillow==9.3.0
